fix typing and remove deprecated comments
